Full Space network example now working
Dijkstra_CGR is really slow for lots of Targets, likely because the CP is so large. No need for it to be, since most of the rows are contacts with targets, which don't need to be in there. We only need the contacts with the target in question. Indeed, this is how it would be if requests were for random locations, rather than pre-defined targets