Working fairly robustly for CGR
Nominal baseline working well, including the rerouting of the second low priority bundle, which gets assigned to the optimal route initially, because linear volume modelling does not account for the delay in arrival of data at later contacts, however it discovers it cannot fit over the final contact, so gets re-routed accordingly.

the larger, high priority bundle is forced to take a longer route due to the fact that it must all arrive at a node before being forwarded from them. As such, despite the optimal route having sufficient "volume", in practice it couldn't make the journey.