Simple MSR vs. CGR test setup
This test demonstrates that MSR avoids the dropping of a bundle by routing around congestion at the downlink. However, since the outbound queues are not ordered, we currently deliver what is (technically) the lower priority bundle, i.e. the one acquired last. Even if this was a higher priority bundle by class, it would still take prescedence as it is ahead of the other bundle in the OBQ.